# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample dating app API.",
        "title": "Dating App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Login with the provided phone number and get OTP.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login Object payload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Payload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP generated successfully",
                        "schema": {
                            "$ref": "#/definitions/main.OTPResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Invalid phone number",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/purchase": {
            "post": {
                "description": "Purchase premium membership.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Purchase premium",
                "parameters": [
                    {
                        "description": "Purchase object",
                        "name": "purchase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Purchase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Purchase successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Register a new user with the provided phone number.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Signup Object payload",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Payload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OTP generated successfully",
                        "schema": {
                            "$ref": "#/definitions/main.OTPResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/swipe": {
            "post": {
                "description": "Perform a swipe action by the user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Perform a swipe action",
                "parameters": [
                    {
                        "description": "Swipe object",
                        "name": "swipe",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Swipe"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Swipe action performed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Get a list of users based on the logged-in user's preferences.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a list of users based on preferences",
                "responses": {
                    "200": {
                        "description": "List of users matching preferences",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.OTPResponse": {
            "type": "object",
            "properties": {
                "otp": {
                    "type": "string"
                }
            }
        },
        "main.Payload": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "phone_number": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "main.Purchase": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "purchase_date": {
                    "type": "string"
                }
            }
        },
        "main.Swipe": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "swipe_date": {
                    "type": "string"
                },
                "swipe_type": {
                    "type": "string"
                }
            }
        },
        "main.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "is_premium": {
                    "type": "boolean"
                },
                "login_at": {
                    "type": "string"
                },
                "logout_at": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "signup_at": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        }
    }
}